requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
Cerberus==1.2
certifi==2018.4.16
chardet==3.0.4
djangorestframework==3.8.2
idna==2.7
more-itertools==4.2.0
pkginfo==1.4.2
pluggy==0.6.0
py==1.5.4
py-testutils==0.6
pytest==3.6.3
requests==2.19.1
requests-toolbelt==0.8.0
six==1.11.0
tqdm==4.23.4
twine==1.11.0
urllib3==1.23
